+++ tests/compiler/generator/spatio-flux/test_spatio_flux_II.py
