+++ tests/compiler/generator/test_toy_add_floats.py
@@ -5,8 +5,8 @@ from typing import Any
 
 import process_bigraph as pg  # type: ignore[import-untyped]
 
-from process_bigraph_lang.compiler.compiler import compile_ast
-from process_bigraph_lang.compiler.generator import assemble_pb
+from process_bigraph_lang.compiler.generator import compile_ast
+from process_bigraph_lang.compiler.converter import assemble_pb
 from process_bigraph_lang.compiler.pb_model import PBStore, PBStep, PBModel, PBProcess
 from process_bigraph_lang.dsl.ast_model import ASTModel
 from process_bigraph_lang.dsl.langium_pblang import langium_parse_pblang_file
